+++ lib/cli/writer.js
@@ -22,10 +22,10 @@ const fs = Bluebird.promisifyAll(require('fs'))
 const mountutils = Bluebird.promisifyAll(require('mountutils'))
 const drivelist = Bluebird.promisifyAll(require('drivelist'))
 const os = require('os')
-const imageStream = require('../image-stream')
+const imageStream = require('../sdk/image-stream')
 const errors = require('../shared/errors')
 const constraints = require('../shared/drive-constraints')
-const ImageWriter = require('../writer')
+const ImageWriter = require('../sdk/writer')
 const diskpart = require('./diskpart')
 
 /**
