+++ lib/gui/app/modules/drive-scanner.js
@@ -21,7 +21,7 @@ const Bluebird = require('bluebird')
 const fs = Bluebird.promisifyAll(require('fs'))
 const path = require('path')
 const settings = require('../models/settings')
-const SDK = require('../../../shared/sdk')
+const SDK = require('../../../sdk')
 
 /**
  * @summary The Etcher "blobs" directory path
@@ -31,7 +31,7 @@ const SDK = require('../../../shared/sdk')
 const BLOBS_DIRECTORY = path.join(__dirname, '..', '..', 'blobs')
 
 const scanner = SDK.createScanner({
-  standard: {
+  blockdevice: {
     get includeSystemDrives () {
       return settings.get('unsafeMode')
     }
