+++ lib/gui/app/pages/main/controllers/image-selection.js
@@ -21,7 +21,7 @@ const Bluebird = require('bluebird')
 const path = require('path')
 const messages = require('../../../../../shared/messages')
 const errors = require('../../../../../shared/errors')
-const imageStream = require('../../../../../image-stream')
+const imageStream = require('../../../../../sdk/image-stream')
 const supportedFormats = require('../../../../../shared/supported-formats')
 const analytics = require('../../../modules/analytics')
 const selectionState = require('../../../../../shared/models/selection-state')
