+++ lib/sdk/adapters/blockdevice/index.js
@@ -22,15 +22,15 @@ const EventEmitter = require('events')
 const drivelist = Bluebird.promisifyAll(require('drivelist'))
 
 /**
- * @summary StandardAdapter
+ * @summary BlockDeviceAdapter
  * @class
  */
-class StandardAdapter extends EventEmitter {
+class BlockDeviceAdapter extends EventEmitter {
   /**
-   * @summary StandardAdapter constructor
+   * @summary BlockDeviceAdapter constructor
    * @class
    * @example
-   * const adapter = new StandardAdapter()
+   * const adapter = new BlockDeviceAdapter()
    */
   constructor () {
     super()
@@ -51,7 +51,7 @@ class StandardAdapter extends EventEmitter {
    * @param {Object} [options] - options
    * @param {Object} [options.includeSystemDrives=false] - include system drives
    * @param {Function} [callback] - optional callback
-   * @returns {StandardAdapter}
+   * @returns {BlockDeviceAdapter}
    *
    * @example
    * adapter.scan({
@@ -101,7 +101,7 @@ class StandardAdapter extends EventEmitter {
  * @type {String}
  * @constant
  */
-StandardAdapter.id = 'standard'
+BlockDeviceAdapter.id = 'blockdevice'
 
 // Exports
-module.exports = StandardAdapter
+module.exports = BlockDeviceAdapter
