+++ lib/sdk/adapters/usbboot/index.js
@@ -27,7 +27,7 @@ const Bluebird = require('bluebird')
 const debug = require('debug')('sdk:usbboot')
 const usb = require('./usb')
 const protocol = require('./protocol')
-const utils = require('../../utils')
+const utils = require('../../../shared/utils')
 
 debug.enabled = true
