+++ lib/sdk/adapters/usbboot/protocol.js
