+++ lib/sdk/adapters/usbboot/usb.js
