+++ lib/sdk/image-stream/archive-hooks/zip.js
@@ -20,7 +20,7 @@ const Bluebird = require('bluebird')
 const _ = require('lodash')
 const StreamZip = require('node-stream-zip')
 const yauzl = Bluebird.promisifyAll(require('yauzl'))
-const errors = require('../../shared/errors')
+const errors = require('../../../shared/errors')
 
 /**
  * @summary Get all archive entries
