+++ lib/sdk/image-stream/archive.js
@@ -21,8 +21,8 @@ const _ = require('lodash')
 const PassThroughStream = require('stream').PassThrough
 const supportedFileTypes = require('./supported')
 const utils = require('./utils')
-const errors = require('../shared/errors')
-const fileExtensions = require('../shared/file-extensions')
+const errors = require('../../shared/errors')
+const fileExtensions = require('../../shared/file-extensions')
 
 /**
  * @summary Archive metadata base path
