+++ lib/sdk/image-stream/gzip.js
