+++ lib/sdk/image-stream/handlers.js
@@ -29,9 +29,9 @@ const udif = Bluebird.promisifyAll(require('udif'))
 const archive = require('./archive')
 const utils = require('./utils')
 const zipArchiveHooks = require('./archive-hooks/zip')
-const fileExtensions = require('../shared/file-extensions')
+const fileExtensions = require('../../shared/file-extensions')
 const path = require('path')
-const errors = require('../shared/errors')
+const errors = require('../../shared/errors')
 
 /**
  * @summary Default image extension to be assumed
