+++ lib/sdk/image-stream/index.js
@@ -23,7 +23,7 @@ const stream = require('stream')
 const mime = require('./mime')
 const handlers = require('./handlers')
 const supportedFileTypes = require('./supported')
-const errors = require('../shared/errors')
+const errors = require('../../shared/errors')
 const parsePartitions = require('./parse-partitions')
 
 /**
@@ -54,7 +54,7 @@ const parsePartitions = require('./parse-partitions')
  * @returns {Promise}
  *
  * @example
- * const imageStream = require('./lib/image-stream');
+ * const imageStream = require('./lib/sdk/image-stream');
  *
  * imageStream.getFromFilePath('path/to/rpi.img.xz').then((image) => {
  *   console.log(`The image display name is: ${image.name}`);
@@ -106,7 +106,7 @@ exports.getFromFilePath = (file) => {
  * @returns {Promise}
  *
  * @example
- * const imageStream = require('./lib/image-stream');
+ * const imageStream = require('./lib/sdk/image-stream');
  *
  * imageStream.getImageMetadata('path/to/rpi.img.xz').then((metadata) => {
  *   console.log(`The image display name is: ${metadata.name}`);
@@ -131,7 +131,7 @@ exports.getImageMetadata = (file) => {
  * @public
  *
  * @example
- * const imageStream = require('./lib/image-stream');
+ * const imageStream = require('./lib/sdk/image-stream');
  *
  * imageStream.supportedFileTypes.forEach((fileType) => {
  *   console.log('Supported file type: ' + fileType.extension);
