+++ lib/sdk/image-stream/mime.js
