+++ lib/sdk/image-stream/parse-partitions.js
