+++ lib/sdk/image-stream/supported.js
