+++ lib/sdk/image-stream/utils.js
@@ -18,7 +18,7 @@
 
 const Bluebird = require('bluebird')
 const fs = Bluebird.promisifyAll(require('fs'))
-const errors = require('../shared/errors')
+const errors = require('../../shared/errors')
 
 /**
  * @summary Read a buffer from an image file descriptor
