+++ lib/sdk/index.js
@@ -1,0 +1,47 @@
+/*
+ * Copyright 2017 resin.io
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+'use strict'
+
+const SDK = module.exports
+
+/**
+ * @summary Initialised adapters
+ * @type {Object<String,Adapter>}
+ * @constant
+ */
+SDK.adapters = require('./adapters')
+
+/**
+ * Adapter Scanner
+ * @see scanner.js
+ * @ignore
+ */
+SDK.Scanner = require('./scanner')
+
+/**
+ * @summary Create a new Scanner
+ * @param {Object} [options] - options
+ * @returns {SDK.Scanner}
+ * @example
+ * SDK.createScanner({
+ *   blockdevice: { ... },
+ *   usbboot: { ... }
+ * })
+ */
+SDK.createScanner = (options) => {
+  return new SDK.Scanner(options)
+}
