+++ lib/sdk/scanner.js
@@ -16,54 +16,28 @@
 
 'use strict'
 
-const EventEmitter = require('events')
 const _ = require('lodash')
-const SDK = module.exports
-const debug = require('debug')('sdk')
-const os = require('os')
+const EventEmitter = require('events')
+const debug = require('debug')('sdk:scanner')
+const SDK = require('./')
 
 debug.enabled = true
-
-/**
- * @summary The list of loaded adapters
- * @type {Object[]}
- * @constant
- */
-const ADAPTERS = [
-  require('./standard')
-]
-
-// We don't support usbboot on GNU/Linux yet, given
-// that some distributions require root permissions
-// to open USB devices.
-if (os.platform() !== 'linux') {
-  ADAPTERS.push(require('./usbboot'))
-}
-
-/**
- * @summary Initialised adapters
- * @type {Object<String,Adapter>}
- * @constant
- */
-SDK.adapters = _.reduce(ADAPTERS, (adapters, Adapter) => {
-  adapters[Adapter.id] = new Adapter()
-  return adapters
-}, {})
 
 /* eslint-disable lodash/prefer-lodash-method */
 
 /**
  * Adapter Scanner
  * @class Scanner
+ * @memberOf SDK
  */
-SDK.Scanner = class Scanner extends EventEmitter {
+class Scanner extends EventEmitter {
   /**
    * @summary Adapter Scanner constructor
    * @param {Object<String,Object>} [options] - device adapter options
    * @param {Object} [options.adapters] - map of external device adapters
    * @example
-   * new SDK.Scanner({
-   *   standard: { ... },
+   * new Scanner({
+   *   blockdevice: { ... },
    *   usbboot: { ... }
    * })
    */
@@ -133,7 +107,7 @@ SDK.Scanner = class Scanner extends EventEmitter {
   /**
    * @summary Start scanning for devices
    * @public
-   * @returns {SDK.Scanner}
+   * @returns {Scanner}
    * @example
    * scanner.start()
    */
@@ -156,7 +130,7 @@ SDK.Scanner = class Scanner extends EventEmitter {
       const runScan = () => {
         adapter.scan(options, () => {
           if (this.isScanning) {
-            setTimeout(runScan, SDK.Scanner.MIN_SCAN_DELAY)
+            setTimeout(runScan, Scanner.MIN_SCAN_DELAY)
           }
         })
       }
@@ -177,7 +151,7 @@ SDK.Scanner = class Scanner extends EventEmitter {
   /**
    * @summary Stop scanning for devices
    * @public
-   * @returns {SDK.Scanner}
+   * @returns {Scanner}
    * @example
    * scanner.stop()
    */
@@ -202,7 +176,7 @@ SDK.Scanner = class Scanner extends EventEmitter {
    * @summary Subscribe to an adapter
    * @public
    * @param {Adapter} adapter - device adapter
-   * @returns {SDK.Scanner}
+   * @returns {Scanner}
    * @example
    * scanner.subscribe(adapter)
    */
@@ -223,7 +197,7 @@ SDK.Scanner = class Scanner extends EventEmitter {
    * @summary Unsubscribe from an adapter
    * @public
    * @param {Adapter} adapter - device adapter
-   * @returns {SDK.Scanner}
+   * @returns {Scanner}
    * @example
    * scanner.unsubscribe(adapter)
    * // OR
@@ -253,18 +227,6 @@ SDK.Scanner = class Scanner extends EventEmitter {
  * @const
  * @type {Number}
  */
-SDK.Scanner.MIN_SCAN_DELAY = 500
+Scanner.MIN_SCAN_DELAY = 500
 
-/**
- * @summary Create a new Scanner
- * @param {Object} [options] - options
- * @returns {SDK.Scanner}
- * @example
- * SDK.createScanner({
- *   standard: { ... },
- *   usbboot: { ... }
- * })
- */
-SDK.createScanner = (options) => {
-  return new SDK.Scanner(options)
-}
+module.exports = Scanner
