+++ lib/sdk/writer/block-read-stream.js
