+++ lib/sdk/writer/block-stream.js
