+++ lib/sdk/writer/block-write-stream.js
