+++ lib/sdk/writer/checksum-stream.js
