+++ lib/sdk/writer/error-types.js
