+++ lib/sdk/writer/index.js
