+++ lib/sdk/writer/progress-stream.js
