+++ lib/shared/supported-formats.js
@@ -18,7 +18,7 @@
 
 const _ = require('lodash')
 const path = require('path')
-const imageStream = require('../image-stream')
+const imageStream = require('../sdk/image-stream')
 const fileExtensions = require('./file-extensions')
 
 /**
