+++ tests/image-stream/img.spec.js
@@ -21,7 +21,7 @@ const fs = require('fs')
 const path = require('path')
 const DATA_PATH = path.join(__dirname, 'data')
 const IMAGES_PATH = path.join(DATA_PATH, 'images')
-const imageStream = require('../../lib/image-stream/index')
+const imageStream = require('../../lib/sdk/image-stream/index')
 const tester = require('./tester')
 
 describe('ImageStream: IMG', function () {
